Issue #530: Major refactor of the download code. Also query/download for SLC almost works now
